Added UserRepository interface and changed code to depend on it. Optimized imports
