+++ RestarogusBackend/src/main/kotlin/org/amogus/restarogus/entities/User.kt
@@ -3,7 +3,7 @@ package org.amogus.restarogus.entities
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.core.userdetails.UserDetails
-import java.util.UUID
+import java.util.*
 
 data class User(
     val id: UUID, val userName: String, val authPassword: String, val role: Role
