+++ RestarogusBackend/src/main/kotlin/org/amogus/restarogus/services/AuthenticationService.kt
@@ -1,11 +1,11 @@
 package org.amogus.restarogus.services
 
-import org.amogus.restarogus.responses.AuthenticationResponse
-import org.amogus.restarogus.requests.LoginRequest
-import org.amogus.restarogus.requests.RegisterRequest
 import org.amogus.restarogus.entities.Role
 import org.amogus.restarogus.entities.User
 import org.amogus.restarogus.repositories.UserRepository
+import org.amogus.restarogus.requests.LoginRequest
+import org.amogus.restarogus.requests.RegisterRequest
+import org.amogus.restarogus.responses.AuthenticationResponse
 import org.amogus.restarogus.services.interfaces.JwtService
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
